Add docs for component with real view-model
This commit will add a new section to the article documentation
that cover the scenario of testing a custom component with a real
view-model, mocking out any dependencies.

I feel that this is a core scenario for any real-world application
that is not well covered in the documentation today, and something
that I struggled with.

This commit is not complete, since I do not know how to number the
new section, and also the wording is a little off, compared to the
rest of the documenation.  But hey, it's a start!